requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0  # Для расчёта индикаторов
python-dotenv>=1.0.0  # Для конфигов
